Renamed Fulfil::Errors to Fulfil::Errors.
