+++ lib/fulfil/error.rb
